Add a cast into a int16 for the workbench mode.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@13412 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/plug.c
@@ -516,7 +516,7 @@ void wstdinit(char *obj, char *parent)
 			if (cid != (int16 *) 0)
 			{
 				dftype = eif_compound_id ((int16 *)0, l[1], 
-										   exp_dtype & EO_TYPE, cid);
+										   (int16) (exp_dtype & EO_TYPE), cid);
 
 				exp_dtype = (exp_dtype & EO_UPPER) | dftype;
 			}
